introduce static buildDelegationPlan function
