+++ packages/delegate/src/mergeFields.ts
@@ -1,152 +1,20 @@
 import {
-  FieldNode,
-  SelectionNode,
-  Kind,
   GraphQLResolveInfo,
   SelectionSetNode,
   GraphQLObjectType,
   responsePathAsArray,
-  getNamedType,
   GraphQLError,
   locatedError,
   GraphQLSchema,
 } from 'graphql';
 
 import { ExternalObject, MergedTypeInfo, SubschemaConfig } from './types';
-import { memoize4, memoize3, memoize2, memoize1 } from './memoize';
+import { memoize4 } from './memoize';
 import { Subschema } from './Subschema';
 import { collectFields, ExecutionContext } from 'graphql/execution/execute.js';
 import { relocatedError } from '@graphql-tools/utils';
 import { FIELD_SUBSCHEMA_MAP_SYMBOL, OBJECT_SUBSCHEMA_SYMBOL, UNPATHED_ERRORS_SYMBOL } from './symbols';
-
-const sortSubschemasByProxiability = memoize4(function sortSubschemasByProxiability(
-  mergedTypeInfo: MergedTypeInfo,
-  sourceSubschemas: Array<Subschema>,
-  targetSubschemas: Array<Subschema>,
-  fieldNodes: Array<FieldNode>
-): {
-  proxiableSubschemas: Array<Subschema>;
-  nonProxiableSubschemas: Array<Subschema>;
-} {
-  // 1.  calculate if possible to delegate to given subschema
-
-  const proxiableSubschemas: Array<Subschema> = [];
-  const nonProxiableSubschemas: Array<Subschema> = [];
-
-  for (const t of targetSubschemas) {
-    const selectionSet = mergedTypeInfo.selectionSets.get(t);
-    const fieldSelectionSets = mergedTypeInfo.fieldSelectionSets.get(t);
-    if (selectionSet != null && !subschemaTypesContainSelectionSet(mergedTypeInfo, sourceSubschemas, selectionSet)) {
-      nonProxiableSubschemas.push(t);
-    } else {
-      if (
-        fieldSelectionSets == null ||
-        fieldNodes.every(fieldNode => {
-          const fieldName = fieldNode.name.value;
-          const fieldSelectionSet = fieldSelectionSets[fieldName];
-          return (
-            fieldSelectionSet == null ||
-            subschemaTypesContainSelectionSet(mergedTypeInfo, sourceSubschemas, fieldSelectionSet)
-          );
-        })
-      ) {
-        proxiableSubschemas.push(t);
-      } else {
-        nonProxiableSubschemas.push(t);
-      }
-    }
-  }
-
-  return {
-    proxiableSubschemas,
-    nonProxiableSubschemas,
-  };
-});
-
-const buildDelegationPlan = memoize3(function buildDelegationPlan(
-  mergedTypeInfo: MergedTypeInfo,
-  fieldNodes: Array<FieldNode>,
-  proxiableSubschemas: Array<Subschema>
-): {
-  delegationMap: Map<Subschema, SelectionSetNode>;
-  unproxiableFieldNodes: Array<FieldNode>;
-} {
-  const { uniqueFields, nonUniqueFields } = mergedTypeInfo;
-  const unproxiableFieldNodes: Array<FieldNode> = [];
-
-  // 2. for each selection:
-
-  const delegationMap: Map<Subschema, SelectionSetNode> = new Map();
-  for (const fieldNode of fieldNodes) {
-    if (fieldNode.name.value === '__typename') {
-      continue;
-    }
-
-    // 2a. use uniqueFields map to assign fields to subschema if one of possible subschemas
-
-    const uniqueSubschema: Subschema = uniqueFields[fieldNode.name.value];
-    if (uniqueSubschema != null) {
-      if (!proxiableSubschemas.includes(uniqueSubschema)) {
-        unproxiableFieldNodes.push(fieldNode);
-        continue;
-      }
-
-      const existingSubschema = delegationMap.get(uniqueSubschema)?.selections as SelectionNode[];
-      if (existingSubschema != null) {
-        existingSubschema.push(fieldNode);
-      } else {
-        delegationMap.set(uniqueSubschema, {
-          kind: Kind.SELECTION_SET,
-          selections: [fieldNode],
-        });
-      }
-
-      continue;
-    }
-
-    // 2b. use nonUniqueFields to assign to a possible subschema,
-    //     preferring one of the subschemas already targets of delegation
-
-    let nonUniqueSubschemas: Array<Subschema> = nonUniqueFields[fieldNode.name.value];
-    if (nonUniqueSubschemas == null) {
-      unproxiableFieldNodes.push(fieldNode);
-      continue;
-    }
-
-    nonUniqueSubschemas = nonUniqueSubschemas.filter(s => proxiableSubschemas.includes(s));
-    if (!nonUniqueSubschemas.length) {
-      unproxiableFieldNodes.push(fieldNode);
-      continue;
-    }
-
-    const existingSubschema = nonUniqueSubschemas.find(s => delegationMap.has(s));
-    if (existingSubschema != null) {
-      // It is okay we previously explicitly check whether the map has the element.
-      (delegationMap.get(existingSubschema)!.selections as SelectionNode[]).push(fieldNode);
-    } else {
-      delegationMap.set(nonUniqueSubschemas[0], {
-        kind: Kind.SELECTION_SET,
-        selections: [fieldNode],
-      });
-    }
-  }
-
-  return {
-    delegationMap,
-    unproxiableFieldNodes,
-  };
-});
-
-const createSubschemas = memoize1(function createSubschemas(sourceSubschema: Subschema): Array<Subschema> {
-  return [sourceSubschema];
-});
-
-const combineSubschemas = memoize2(function combineSubschemas(
-  sourceSubschemas: Array<Subschema>,
-  additionalSubschemas: Array<Subschema>
-): Array<Subschema> {
-  return sourceSubschemas.concat(additionalSubschemas);
-});
+import { buildDelegationPlan } from './buildDelegationPlan';
 
 export function isExternalObject(data: any): data is ExternalObject {
   return data[UNPATHED_ERRORS_SYMBOL] !== undefined;
@@ -175,54 +43,29 @@ export function getUnpathedErrors(object: ExternalObject): Array<GraphQLError> {
 
 export async function mergeFields(
   mergedTypeInfo: MergedTypeInfo,
-  typeName: string,
   object: any,
-  fieldNodes: Array<FieldNode>,
   sourceSubschema: Subschema<any, any, any, any>,
   targetSubschemas: Array<Subschema<any, any, any, any>>,
   context: any,
   info: GraphQLResolveInfo
 ): Promise<any> {
-  return _mergeFields(
-    mergedTypeInfo,
-    typeName,
-    object,
-    fieldNodes,
-    createSubschemas(sourceSubschema),
-    targetSubschemas,
-    context,
-    info
-  );
+  const delegationMaps = buildDelegationPlanFromInfo(info, mergedTypeInfo, sourceSubschema, targetSubschemas);
+
+  for (const delegationMap of delegationMaps) {
+    object = await executeDelegationStage(mergedTypeInfo, delegationMap, object, context, info);
+  }
+
+  return object;
 }
 
-async function _mergeFields(
+async function executeDelegationStage(
   mergedTypeInfo: MergedTypeInfo,
-  typeName: string,
-  object: any,
-  fieldNodes: Array<FieldNode>,
-  sourceSubschemas: Array<Subschema<any, any, any, any>>,
-  targetSubschemas: Array<Subschema<any, any, any, any>>,
+  delegationMap: Map<Subschema, SelectionSetNode>,
+  object: ExternalObject,
   context: any,
   info: GraphQLResolveInfo
 ): Promise<any> {
-  if (!fieldNodes.length) {
-    return object;
-  }
-
-  const { proxiableSubschemas, nonProxiableSubschemas } = sortSubschemasByProxiability(
-    mergedTypeInfo,
-    sourceSubschemas,
-    targetSubschemas,
-    fieldNodes
-  );
-
-  const { delegationMap, unproxiableFieldNodes } = buildDelegationPlan(mergedTypeInfo, fieldNodes, proxiableSubschemas);
-
-  if (!delegationMap.size) {
-    return object;
-  }
-
-  const combinedErrors = object[UNPATHED_ERRORS_SYMBOL] || [];
+  const combinedErrors = object[UNPATHED_ERRORS_SYMBOL] ?? [];
 
   const path = responsePathAsArray(info.path);
 
@@ -288,51 +131,25 @@ async function _mergeFields(
 
   combinedResult[UNPATHED_ERRORS_SYMBOL] = combinedErrors;
 
-  return _mergeFields(
-    mergedTypeInfo,
-    typeName,
-    combinedResult,
-    unproxiableFieldNodes,
-    combineSubschemas(sourceSubschemas, proxiableSubschemas),
-    nonProxiableSubschemas,
-    context,
-    info
-  );
+  return combinedResult;
 }
 
-const subschemaTypesContainSelectionSet = memoize3(function subschemaTypesContainSelectionSetMemoized(
+const buildDelegationPlanFromInfo = memoize4(function buildDelegationPlanFromInfo(
+  info: GraphQLResolveInfo,
   mergedTypeInfo: MergedTypeInfo,
-  sourceSubchemas: Array<Subschema>,
-  selectionSet: SelectionSetNode
-) {
-  return typesContainSelectionSet(
-    sourceSubchemas.map(
-      sourceSubschema => sourceSubschema.transformedSchema.getType(mergedTypeInfo.typeName) as GraphQLObjectType
-    ),
-    selectionSet
+  sourceSubschema: Subschema<any, any, any, any>,
+  targetSubschemas: Array<Subschema<any, any, any, any>>
+): Array<Map<Subschema, SelectionSetNode>> {
+  const { schema, fragments, variableValues, fieldNodes } = info;
+  const stitchingInfo = schema.extensions?.['stitchingInfo'];
+  return buildDelegationPlan(
+    schema,
+    stitchingInfo,
+    mergedTypeInfo,
+    sourceSubschema,
+    targetSubschemas,
+    fieldNodes,
+    fragments,
+    variableValues
   );
 });
-
-function typesContainSelectionSet(types: Array<GraphQLObjectType>, selectionSet: SelectionSetNode): boolean {
-  const fieldMaps = types.map(type => type.getFields());
-
-  for (const selection of selectionSet.selections) {
-    if (selection.kind === Kind.FIELD) {
-      const fields = fieldMaps.map(fieldMap => fieldMap[selection.name.value]).filter(field => field != null);
-      if (!fields.length) {
-        return false;
-      }
-
-      if (selection.selectionSet != null) {
-        return typesContainSelectionSet(
-          fields.map(field => getNamedType(field.type)) as Array<GraphQLObjectType>,
-          selection.selectionSet
-        );
-      }
-    } else if (selection.kind === Kind.INLINE_FRAGMENT && selection.typeCondition?.name.value === types[0].name) {
-      return typesContainSelectionSet(types, selection.selectionSet);
-    }
-  }
-
-  return true;
-}
